+++ RTBEngine/Engine/RTBEngine.h
@@ -13,7 +13,9 @@
 #include "Scene/ComponentQuery.h"
 #include "ECS/Entity.h"
 #include "ECS/World.h"
+#include "ECS/EcsStats.h"
 #include "ECS/ProjectileSimulation.h"
+#include "ECS/SwarmSimulation.h"
 #include "Scripting/LatentActions.h"
 
 // Scene & Components